refactor : uploadAll() 메소드 생성(반복적인 저장)
