+++ src/main/java/com/samcomo/dbz/report/service/impl/ReportServiceImpl.java
@@ -1,8 +1,6 @@
 package com.samcomo.dbz.report.service.impl;
 
 import com.samcomo.dbz.global.exception.ErrorCode;
-import com.samcomo.dbz.global.s3.ImageType;
-import com.samcomo.dbz.global.s3.ImageUploadState;
 import com.samcomo.dbz.global.s3.S3Service;
 import com.samcomo.dbz.member.exception.MemberException;
 import com.samcomo.dbz.member.model.entity.Member;
@@ -13,7 +11,6 @@ import com.samcomo.dbz.report.model.dto.ReportDto;
 import com.samcomo.dbz.report.model.dto.ReportImageDto;
 import com.samcomo.dbz.report.model.dto.ReportList;
 import com.samcomo.dbz.report.model.dto.ReportStateDto;
-import com.samcomo.dbz.report.model.dto.ReportStateDto.Response;
 import com.samcomo.dbz.report.model.entity.Report;
 import com.samcomo.dbz.report.model.entity.ReportImage;
 import com.samcomo.dbz.report.model.repository.ReportImageRepository;
@@ -23,6 +20,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Repository;
@@ -30,6 +28,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 @Repository
 @RequiredArgsConstructor
+@Slf4j
 public class ReportServiceImpl implements ReportService {
 
   private final ReportRepository reportRepository;
@@ -46,25 +45,7 @@ public class ReportServiceImpl implements ReportService {
         .orElseThrow(() -> new MemberException(ErrorCode.MEMBER_NOT_FOUND));
 
     // S3 이미지 저장
-    List<ReportImage> imageList = new ArrayList<>();
-    if (!multipartFileList.isEmpty()) {
-      for (MultipartFile image : multipartFileList) {
-        ImageUploadState imageUploadState = s3Service.upload(image, ImageType.REPORT);
-
-        // 이미지 업로드 실패
-        if(!imageUploadState.isSuccess()){
-          //지금까지 저장된 이미지 삭제
-          deleteUploadedImages(imageList);
-
-          throw new ReportException(ErrorCode.IMAGE_UPLOAD_FAIL);
-        }
-
-        String imageUrl = imageUploadState.getImageUrl();
-        imageList.add(ReportImage.builder()
-            .imageUrl(imageUrl)
-            .build());
-      }
-    }
+    List<ReportImage> imageList = s3Service.uploadAll(multipartFileList);
 
     // 게시글 저장
     Report newReport = reportRepository.save(Report.from(reportForm, member));
@@ -123,7 +104,7 @@ public class ReportServiceImpl implements ReportService {
 
   @Override
   public ReportDto.Response updateReport(
-      long reportId, ReportDto.Form reportForm, List<MultipartFile> imageList, long userId
+      long reportId, ReportDto.Form reportForm, List<MultipartFile> multipartFileList, long userId
   ) {
 
     Member member = memberRepository.findById(userId)
@@ -157,23 +138,23 @@ public class ReportServiceImpl implements ReportService {
     }
 
     // 변경된 이미지 저장
-    List<ReportImage> newImageList = new ArrayList<>();
-    for (MultipartFile image : imageList) {
-      ImageUploadState imageUploadState = s3Service.upload(image, ImageType.REPORT);
-
-      // 이미지 업로드 실패
-      if(!imageUploadState.isSuccess()){
-        //지금까지 저장된 이미지 삭제
-        deleteUploadedImages(newImageList);
-
-        throw new ReportException(ErrorCode.IMAGE_UPLOAD_FAIL);
-      }
-
-      String imageUrl = imageUploadState.getImageUrl();
-      newImageList.add(ReportImage.builder()
-          .imageUrl(imageUrl)
-          .build());
-    }
+    List<ReportImage> newImageList = s3Service.uploadAll(multipartFileList);
+//    for (MultipartFile image : imageList) {
+//      ImageUploadState imageUploadState = s3Service.upload(image, ImageType.REPORT);
+//
+//      // 이미지 업로드 실패
+//      if(!imageUploadState.isSuccess()){
+//        //지금까지 저장된 이미지 삭제
+//        deleteUploadedImages(newImageList);
+//
+//        throw new ReportException(ErrorCode.IMAGE_UPLOAD_FAIL);
+//      }
+//
+//      String imageUrl = imageUploadState.getImageUrl();
+//      newImageList.add(ReportImage.builder()
+//          .imageUrl(imageUrl)
+//          .build());
+//    }
 
     // 게시글 저장
     Report newReport = reportRepository.save(report);
@@ -190,7 +171,7 @@ public class ReportServiceImpl implements ReportService {
   }
 
   @Override
-  public Response deleteReport(long userId, long reportId) {
+  public ReportStateDto.Response deleteReport(long userId, long reportId) {
 
     Member member = memberRepository.findById(userId)
         .orElseThrow(() -> new MemberException(ErrorCode.MEMBER_NOT_FOUND));
@@ -213,12 +194,12 @@ public class ReportServiceImpl implements ReportService {
 
     return ReportStateDto.Response.builder()
         .reportId(report.getId())
-        .status(ReportStatus.DELETED.getDescription())
+        .status(ReportStatus.DELETED.toString())
         .build();
   }
 
   @Override
-  public Response changeStatusToFound(long userId, long reportId) {
+  public ReportStateDto.Response changeStatusToFound(long userId, long reportId) {
 
     Member member = memberRepository.findById(userId)
         .orElseThrow(() -> new MemberException(ErrorCode.MEMBER_NOT_FOUND));
@@ -235,12 +216,12 @@ public class ReportServiceImpl implements ReportService {
 
     return  ReportStateDto.Response.builder()
         .reportId(savedReport.getId())
-        .status(ReportStatus.FOUND.getDescription())
+        .status(ReportStatus.FOUND.toString())
         .build();
   }
 
   @Override
-  public Slice<ReportList> search(String object, boolean showsInProgressOnly, Pageable pageable) {
+  public Slice<ReportList> searchReport(String object, boolean showsInProgressOnly, Pageable pageable) {
 
     Slice<Report> reportSlice = showsInProgressOnly ?
         reportRepository
@@ -255,7 +236,6 @@ public class ReportServiceImpl implements ReportService {
   private Slice<ReportList> getReportListSlice(Slice<Report> reportSlice) {
 
     return reportSlice.map(report -> {
-
       ReportList reportList = ReportList.from(report);
 
       reportImageRepository.findFirstByReport(report)
@@ -265,18 +245,12 @@ public class ReportServiceImpl implements ReportService {
     });
   }
 
-
-  private boolean checkImageUpload(ImageUploadState imageUploadState) {
-
-    return imageUploadState.isSuccess();
-  }
-
-  private void deleteUploadedImages(List<ReportImage> imageList){
-    for (ReportImage reportImage : imageList){
-      String imageUrl = reportImage.getImageUrl();
-      int idx = imageUrl.lastIndexOf("/");
-      String fileName = imageUrl.substring(idx + 1);
-      s3Service.delete(fileName);
-    }
-  }
+//  private void deleteUploadedImages(List<ReportImage> imageList){
+//    for (ReportImage reportImage : imageList){
+//      String imageUrl = reportImage.getImageUrl();
+//      int idx = imageUrl.lastIndexOf("/");
+//      String fileName = imageUrl.substring(idx + 1);
+//      s3Service.delete(fileName);
+//    }
+//  }
 }
